requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
openai==1.95.1
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pillow==11.3.0
propcache==0.3.2
prov==2.1.1
puremagic==1.30
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
PyJWT==2.10.1
PyMuPDF==1.26.3
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.2
twilio==9.6.5
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whisper==1.1.10
yarl==1.20.1
